+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 =======
 
+Unreleased
+----------
+- add integrated care review CRF (Form 26), required at 6m and 12m visits
+
 0.1.67
 ------
 - bump to edc 0.3.18